config: extract ALG compilation to compiler_security_alg.go
Part B of #4121 (pure code motion). Moves the final cluster —
compileALG (dns/ftp/sip/tftp disable parsing) — verbatim out of
compiler_security.go into its own module.

After this commit compiler_security.go contains only the compileSecurity
dispatcher (which routes each `security { ... }` child to the extracted
per-concern compilers) plus its single fmt import. The 2357-line
grab-bag is now eight focused files, all in package config:

  compiler_security.go             compileSecurity dispatcher
  compiler_security_zones.go       zone compilation
  compiler_security_policy.go      policy compilation
  compiler_security_screen.go      screen/IDS compilation
  compiler_security_addressbook.go address-book + zone-local resolution
  compiler_security_log.go         security-log compilation
  compiler_security_flow.go        flow + tcp-mss compilation
  compiler_security_alg.go         ALG compilation

Function bodies are byte-identical to origin/master; the declaration
stream reconstructed from the eight files exactly equals the original
(verified: every one of the 45 declarations appears verbatim in exactly
one file). compileALG uses no std imports, so the new file has no
import block.

Validation: gofmt -l clean, go vet ./pkg/config/, go build ./...,
go test ./pkg/config/... all green.
