+++ pkg/config/compiler_security.go
@@ -94,27 +94,3 @@ func compileSecurity(node *Node, sec *SecurityConfig) error {
 	}
 	return nil
 }
-
-func compileALG(node *Node, sec *SecurityConfig) error {
-	if dnsNode := node.FindChild("dns"); dnsNode != nil {
-		if dnsNode.FindChild("disable") != nil {
-			sec.ALG.DNSDisable = true
-		}
-	}
-	if ftpNode := node.FindChild("ftp"); ftpNode != nil {
-		if ftpNode.FindChild("disable") != nil {
-			sec.ALG.FTPDisable = true
-		}
-	}
-	if sipNode := node.FindChild("sip"); sipNode != nil {
-		if sipNode.FindChild("disable") != nil {
-			sec.ALG.SIPDisable = true
-		}
-	}
-	if tftpNode := node.FindChild("tftp"); tftpNode != nil {
-		if tftpNode.FindChild("disable") != nil {
-			sec.ALG.TFTPDisable = true
-		}
-	}
-	return nil
-}
